size

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of char: %ld byte(s)\n", sizeof(a));
-	printf("Size of int: %ld byte(s)\n", sizeof(b));
-	printf("Size of long int: %ld byte(s)\n", sizeof(c));
-	printf("Size of long long int: %ld byte(s)\n", sizeof(d));
-	printf("Size of float: %ld byte(s)\n", sizeof(e));
+	printf("Size of char: %ld byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of int: %ld byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of long int: %ld byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of long long int: %ld byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of float: %ld byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
